Remove git status from Makefile
The git status command was used to detect which compressed files were modified.
Now that the compressed files are no longer stored in the repository, this is
no longer needed.

It also complicates the built process as older versions of git return 1 (non-
zero) exit status which prevents later commands from executing.