Fix nav and header text rendering in webkit upon animation.
Webkit (OSX): when a CSS animation fueled by hardware acceleration on the page is triggered, webkit adjusts the fonts from vector to texture/image during the animation. This is particularly noticeable on bolder fonts with text-shadow, such as the primary navigation. In this case, the font switches to normal weight and then back to bold.

This fix could be used on something more global, however doing so has implications for main content areas.

* http://stackoverflow.com/questions/10218039/css3-transition-messing-up-fonts-in-webkit
* http://css-tricks.com/beefing-up-dull-text-in-webkit/