Add size specification to h6 headings
Without a size specified for h6, the browser renders the text as 16px which is much
smaller than the body font. This increases the size of the h5 to 1.111 and sets the
h6 size as 1rem.

The h6 header is used in books within the CMS as well as other documents with deep
hierarchy.