Prevent timeouts
With all tests passing, this test case fails due to a timeout.  This is because it takes more than 60 seconds to run, and there is no output if the tests do not fail.  The work around is to add output for every test.  Pear runs the test in another process, then gets output via a stream.  If there is no output from the stream in 60 seconds, it times out.