Fix the algorithm of allowing UNL Search URL extending
Local sites should only be allowed to append the U or CX querystring
params for use in the WDN search tool.

Logic has been added to enforce the use of ONLY the UNL Search app from
the WDN search tool.

This commit also removes the experimental speak-to-text attribute that
has been removed from the only browser that supported it.

The undocumented feature to add a local placeholder has also been
removed (unapproved by the style guide)