increase swipe length required to open navigation
When scrolling down a long page of information on a touch enabled mobile device, it was easy to accidentally open the side navigation.  This was frustrating and not expected.  This commit aims to address that issue by doubling the distance required to swipe for the navigation to open.  While it might not be a perfect solution, it is hopfully better than it used to be.