Add non-breaking spaces to our link in footer.
It's been bothering me for a while to look at this, and on large widths our name is split into two lines. As the main purveyor of web material, we should probably consolidate that into one line. Open to discussion on it though. There are no ill effects at smaller widths as the font size decreases enough to make it not an issue.