Rewrite build system for Grunt
Replaces various shell, perl/php, and Makefile scripts with Grunt or
NodeJS scripts that handle all of the build. The dependencies for
building have been greatly reduced by consolidating to a single tool.
The README has been updated to reflect this change.

A side-effect of this rewrite is an improvement to the smudge/clean
keyword filters. Regardless of where you build from, the version number
keywords should always be replaced consistently.