Fix a11y errors from pa11y
make the background-color explicit, so that phantomjs will fall back to just using the solid color. If this doesn't happen, phantomjs will not apply a background to the footer at all, thus making all the text in the footer fail contrast checks. I'm not sure of the root cause, but this is the only way that I can get this to work right now.