Fix login link race condition
It looks like there was a race condition. Website can set the login URL with JS, which will also cause `displayLogin()` to be called. `displayLogin()` would remove the ``.loggedin` class from IDM container (but not remove the profile image, etc).  This would cause weird styling.

If a user is already logged in, the `displayLogin()` method should never be called, so prevent it from doing anything.