hide icon fonts from screen readers
This is not currently an issue because in my testing, screen readers have not tried to read these characters. However, nothing is preventing them from doing so in the future (however unlikely). More to the point: this brings the framework in line with what we are recommending in the main content area.