Disable The direction nav by default
The direction nav fails accessibility audits by default because in its 'hidden' state, the navigation is still in the DOM and focusable. However, when focused with a keyboard, they are not shown, and thus can cause confusion (where is the current focus?). Additionally, the animation to show them on hover floats them in from the sides and can create a confusing and moving click target. Let's just disable them for now.