Force events widgets to use https to retrieve events
Safari, mobile Safari, and occasionally IE/Edge (some versions)
would not correctly upgrade requests to the UNL Events API
to https. This would cause a CSRF warning and the request would
fail, causing the events widget not to appear. The problem was
fixed by the user entering https:// before their calendar URL
as opposed to just //, but this patch forces https:// if the
user had the old way put in.

Fixes card
https://trello.com/c/7c3VGIvp/60-events-widget-force-https-urls-for-parameters-in-the-widget