Search form input field max-width
There may be instances when the identity management widget is not present. The search form then grows to fill the available space. This results in a comically wide search input at wider browser widths.

This commit applies a max-width to the nav-full search form input field and right-aligns the search form contents with ‘justify-content: flex-end’ (located in the CSS with rules that change the flex-direction from ‘column’ to ‘row’ at @bp480).