Add delay to deleting Critical CSS when all is cached
When doing a hard refresh on speedy template, there
is a flash of NO CSS. This occurs because the javascript
quickly throws in the deferred CSS to be loaded from cache,
then immediately deletes the critical CSS, which leaves no CSS
until the deferred CSS is loaded.

This change adds a 1 second delay before removing this CSS. This
should be enough time for almost all computers and phones to get
their stylesheet, parse, and render it to avoid this flash of no CSS.

I looked at simply not deleting the critical CSS, but that was a bigger
change that didn't need to be made at this time and without understanding
everything that is placed in each part, it's probably not a great
idea to tinker with it at this time.