Rename ‘.wdn-icon-container’ class
WebAudit expects framework icon classes (‘.wdn-icon-*’) to contain the aria-hidden=“true” attribute for accessibility. However, ‘.wdn-icon-container’ is not an icon itself; instead it can be applied to the parent element of an icon to prevent adjacent text from wrapping below it on a second line, treating icons like hanging indents.