Fix search params on desktop
Setting the url to search (`u` parameter) was not working at desktop widths. This appears to be due to the fact that the search form is never actually submitted on desktop. Instead, a frame is created via js, then cross frame communication is used to submit changes. Only the search query value is passed via the cross frame communication.

This fixes the issue by altering the initial search frame URL that used to include the parameters.

An example of such a configuration is:

```
<link rel="search" href="https://search.unl.edu/?u=https://admissions.unl.edu/" />
```