Do not apply deprecated styles to .dcf-table
Currently, the deprecated styles apply borders and striped backgrounds to all tables. The DCF tables do not add these by default, but instead provides them as optional classes that can be used separately or combined together. This changes the deprecated CSS to target only tables not using the `.dcf-table` class. If your markup is old and not yet using DCF for tables, you’ll get stripes and borders, same as before. If you’ve started using DCF for tables, you’ll need to use the provided classes to add stripes (`.dcf-table-striped`) and/or borders (`.dcf-table-bordered`) as needed.